requirements.txt
tqdm
pandas
scipy
opencv-python
pillow